add manage order- search- list- detail

+++ src/main/java/com/ecom2/customer/dto/CustomerAddressDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CustomerAddressDTO {
-    private int addressId;
+    private long addressId;
     private String address;
     private String commune;
     private String district;
